chore: split schema name tests and properly serialize cast
