+++ tests/common/features/mod.rs
@@ -26,7 +26,6 @@ pub mod pi;
 pub mod repository;
 pub mod satellite;
 pub mod schema;
-pub mod schema_enum;
 pub mod sea_orm_active_enums;
 pub mod self_join;
 pub mod teas;
@@ -58,7 +57,6 @@ pub use metadata::Entity as Metadata;
 pub use repository::Entity as Repository;
 pub use satellite::Entity as Satellite;
 pub use schema::*;
-pub use schema_enum::Entity as SchemaEnum;
 pub use sea_orm_active_enums::*;
 pub use self_join::Entity as SelfJoin;
 pub use teas::Entity as Teas;
